added registrations dashboard
