+++ src/Krg.Website/Areas/Admin/Views/Events/Index.cshtml
@@ -1,0 +1,1 @@
+<h1>This is events view</h1>
