+++ src/Krg.Website/Areas/Admin/Views/Shared/_Layout.cshtml
@@ -19,12 +19,24 @@
     <div class="container-fluid">
         <div class="row">
             <div class="col-sm-auto sticky-top text-bg-dark">
-                <div class="d-flex flex-sm-column flex-row flex-nowrap text-decoration-none align-items-center sticky-top">
-                    <a href="/Admin" class="d-block p-3 text-white text-decoration-none" title="" data-bs-toggle="tooltip" data-bs-placement="right" data-bs-original-title="Icon-only">
+                <div class="d-flex flex-sm-column flex-row flex-nowrap text-decoration-none align-items-start sticky-top">
+                    <a href="/Admin" class="d-block py-3 px-2 text-white text-decoration-none" title="" data-bs-toggle="tooltip" data-bs-placement="right" data-bs-original-title="Icon-only">
                         <i class="bi-house fs-4 me-2"></i>
                         <span class="fs-4">Home</span>
                     </a>
-                    <ul class="nav nav-pills nav-flush flex-sm-column flex-row flex-nowrap mb-auto mx-auto text-center align-items-center">                        
+                    <ul class="nav nav-pills nav-flush flex-sm-column flex-row flex-nowrap mb-auto mx-auto text-center align-items-start">
+                        <li>
+                            <a href="/admin/events" class="nav-link py-3 px-2 text-white" title="" data-bs-toggle="tooltip" data-bs-placement="right" data-bs-original-title="Stats">
+                                <i class="bi-calendar4-event fs-4 me-2"></i>
+                                <span class="fs-4">Events</span>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="/admin/registrations" class="nav-link py-3 px-2 text-white" title="" data-bs-toggle="tooltip" data-bs-placement="right" data-bs-original-title="Stats">
+                                <i class="bi-clipboard-check fs-4 me-2"></i>
+                                <span class="fs-4">Registrations</span>
+                            </a>
+                        </li>
                         <li>
                             <a href="/admin/stats" class="nav-link py-3 px-2 text-white" title="" data-bs-toggle="tooltip" data-bs-placement="right" data-bs-original-title="Stats">
                                 <i class="bi-speedometer2 fs-4 me-2"></i>
@@ -53,7 +65,8 @@
 
 
 
-	<!-- Scripts -->
+    <!-- Scripts -->
+    <script src="~/lib/jquery/dist/jquery.min.js"></script>
 	<script src="https://www.unpkg.com/aspnet-client-validation/dist/aspnet-validation.min.js" crossorigin="anonymous"></script>
 	<script>
 		const v = new aspnetValidation.ValidationService()
