+++ src/Krg.Website/Program.cs
@@ -63,13 +63,13 @@ try
 	app.UseAuthorization();
 
 	app.MapControllerRoute(
+		name: "Admin",
+		pattern: "{area:exists}/{controller=Home}/{action=Index}/{id?}"
+	);
+	
+	app.MapControllerRoute(
 		name: "default",
 		pattern: "{controller=Home}/{action=Index}/{id?}");
-
-    app.MapControllerRoute(
-        name: "Admin",
-        pattern: "{area:exists}/{controller=Home}/{action=Index}/{id?}"
-    );
 
     app.Run();
